style: muda estilo de disposição das pastas localizadas nas pastas components e pages
