+++ src/pages/Login/_login.scss
@@ -1,7 +1,3 @@
-@import '../css/colors';
-
-@import '../css/functions';
-
 .container-form {
   align-items: center;
   background-color: $sweet-dark;
