+++ src/pages/Recipes/Drinks/FiltersRecipesDrinks.jsx
@@ -1,8 +1,8 @@
 import { func, string } from 'prop-types';
 import React, { useEffect, useState } from 'react';
 import { connect } from 'react-redux';
-import { fetchRecipesForCategory } from '../../redux/actions';
-import { fetchCategoriesDrinksApi } from '../../services/fetchApi';
+import { fetchRecipesForCategory } from '../../../redux/actions';
+import { fetchCategoriesDrinksApi } from '../../../services/fetchApi';
 
 function FiltersRecipesDrinks({ getCategory, category }) {
   const [categories, setCategories] = useState([]);
