+++ src/pages/Recipes/Foods/FoodRecipeCards.jsx
@@ -1,23 +1,24 @@
 import { arrayOf, shape } from 'prop-types';
 import React from 'react';
 import { connect } from 'react-redux';
-import RecipeCard from '../../components/RecipeCard';
+import RecipeCard from '../../../components/RecipeCard';
 
 const ZERO = 0;
 const TWELVE = 12;
 
-function DrinkRecipeCards({ recipes }) {
+function FoodRecipeCards({ recipes }) {
   return (
     <section className="main-recipes">
       {recipes.slice(ZERO, TWELVE)
         .map((recipe, index) => (
           <RecipeCard
-            id={ recipe.idDrink }
+            foodPage
+            id={ recipe.idMeal }
             key={ index }
-            name={ recipe.strDrink }
-            src={ recipe.strDrinkThumb }
+            name={ recipe.strMeal }
+            src={ recipe.strMealThumb }
             index={ index }
-            alt={ `${recipe.strDrink} image` }
+            alt={ `${recipe.strMeal} image` }
           />
         ))}
     </section>
@@ -28,8 +29,8 @@ const mapStateToProps = (state) => ({
   recipes: state.recipesReducer.recipes,
 });
 
-DrinkRecipeCards.propTypes = {
+FoodRecipeCards.propTypes = {
   recipes: arrayOf(shape()).isRequired,
 };
 
-export default connect(mapStateToProps)(DrinkRecipeCards);
+export default connect(mapStateToProps)(FoodRecipeCards);
