+++ src/pages/Recipes/_detailsRecipe.scss
