+++ src/pages/explore/_exploreByIngredients.scss
