--- src/pages/myRecipes/_global.scss
@@ -1,5 +1,0 @@
-@import './FavoritesRecipes/FavoriteRecipesCard/favoriteRecipesCard';
-
-@import './MadeRecipes/MadeRecipesCard/madeRecipesCard';
-
-@import './HandleFavoriteAndShare/handleFavoriteAndShare';
