--- src/pages/styles/_global.scss
@@ -1,15 +1,0 @@
-@import './recipes';
-
-@import './detailsRecipe';
-
-@import 'progressRecipe';
-
-@import './myRecipes';
-
-@import './profile';
-
-@import './explore';
-
-@import './exploreByIngredients';
-
-@import './exploreFoodsByArea';
