--- src/pages/styles/notFound.css
@@ -1,9 +1,0 @@
-.not-found {
-  align-items: center;
-  display: flex;
-  font-size: xxx-large;
-  font-weight: 700;
-  height: 100vh;
-  justify-content: center;
-  width: 100vw;
-}
